requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
numpy==1.18.5
pandas==1.0.4
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2020.1
pyxlsb==1.0.6
six==1.15.0
SQLAlchemy==1.3.17
xlrd==1.2.0